Update explore_exploit.py
Replaced the section on "Did we pick the right arm?" --> fixed the code, before we were looking at the one with the highest value, but that can skew the results! We are looking for the row that we picked the most, and that's the one with the most amount of values that !=0... it's a cool bit of code that's worth remembering.